Relocate spring3 and spring5 implementations to a different package
