+++ blueprint-integration/blueprint-integration-spring-5/build.gradle
