+++ blueprint-integration/blueprint-integration-spring-5/src/main/java/com/github/dynamicextensionsalfresco/blueprint/spring5/Spring5HostApplicationContext.java
@@ -1,4 +1,4 @@
-package com.github.dynamicextensionsalfresco.blueprint;
+package com.github.dynamicextensionsalfresco.blueprint.spring5;
 
 import java.io.IOException;
 import java.lang.annotation.Annotation;
@@ -34,7 +34,7 @@ import org.springframework.util.Assert;
  * @author Laurens Fridael
  * @author Laurent Van der Linden
  */
-class Spring5HostApplicationContext extends AbstractApplicationContext {
+public class Spring5HostApplicationContext extends AbstractApplicationContext {
 
 	private final ApplicationContext applicationContext;
 
@@ -220,7 +220,7 @@ class Spring5HostApplicationContext extends AbstractApplicationContext {
 		return applicationContext.getAliases(name);
 	}
 
-	Spring5HostApplicationContext(final ApplicationContext applicationContext) {
+	public Spring5HostApplicationContext(final ApplicationContext applicationContext) {
 		Assert.notNull(applicationContext);
 		this.applicationContext = applicationContext;
 	}
