+++ blueprint-integration/blueprint-integration-spring-5/src/test/java/com/github/dynamicextensionsalfresco/blueprint/Spring5OsgiAutowireBeanFactoryTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNotSame;
 import static org.junit.Assert.assertSame;
 
+import com.github.dynamicextensionsalfresco.blueprint.spring5.Spring5OsgiAutowireBeanFactory;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.context.support.ClassPathXmlApplicationContext;
